Change since parameter type

+++ src/main/java/org/mule/twitter/TwitterConnector.java
@@ -297,7 +297,7 @@ public class TwitterConnector implements MuleContextAware {
     @Processor
     public ResponseList<Status> getHomeTimeline(@Placement(group = "Pagination") @Default(value = "1") @Optional int page,
                                                 @Placement(group = "Pagination") @Default(value = "100") @Optional int count,
-                                                @Placement(group = "Pagination") @Default(value = "-1") @Optional int sinceId)
+                                                @Placement(group = "Pagination") @Default(value = "-1") @Optional long sinceId)
             throws TwitterException {
         return twitter.getHomeTimeline(getPaging(page, count, sinceId));
     }
@@ -333,7 +333,7 @@ public class TwitterConnector implements MuleContextAware {
     public ResponseList<Status> getUserTimelineByScreenName(String screenName,
                                                             @Placement(group = "Pagination") @Default(value = "1") @Optional int page,
                                                             @Placement(group = "Pagination") @Default(value = "100") @Optional int count,
-                                                            @Placement(group = "Pagination") @Default(value = "-1") @Optional int sinceId)
+                                                            @Placement(group = "Pagination") @Default(value = "-1") @Optional long sinceId)
             throws TwitterException {
         return twitter.getUserTimeline(screenName, getPaging(page, count, sinceId));
     }
@@ -369,7 +369,7 @@ public class TwitterConnector implements MuleContextAware {
     public ResponseList<Status> getUserTimelineByUserId(long userId,
                                                         @Placement(group = "Pagination") @Default(value = "1") @Optional int page,
                                                         @Placement(group = "Pagination") @Default(value = "100") @Optional int count,
-                                                        @Placement(group = "Pagination") @Default(value = "-1") @Optional int sinceId)
+                                                        @Placement(group = "Pagination") @Default(value = "-1") @Optional long sinceId)
             throws TwitterException {
         return twitter.getUserTimeline(userId, getPaging(page, count, sinceId));
     }
@@ -411,7 +411,7 @@ public class TwitterConnector implements MuleContextAware {
     @Processor
     public ResponseList<Status> getUserTimeline(@Placement(group = "Pagination") @Default(value = "1") @Optional int page,
                                                 @Placement(group = "Pagination") @Default(value = "100") @Optional int count,
-                                                @Placement(group = "Pagination") @Default(value = "-1") @Optional int sinceId)
+                                                @Placement(group = "Pagination") @Default(value = "-1") @Optional long sinceId)
             throws TwitterException {
         return twitter.getUserTimeline(getPaging(page, count, sinceId));
     }
@@ -438,7 +438,7 @@ public class TwitterConnector implements MuleContextAware {
     @Processor
     public ResponseList<Status> getMentions(@Placement(group = "Pagination") @Default(value = "1") @Optional int page,
                                             @Placement(group = "Pagination") @Default(value = "100") @Optional int count,
-                                            @Placement(group = "Pagination") @Default(value = "-1") @Optional int sinceId)
+                                            @Placement(group = "Pagination") @Default(value = "-1") @Optional long sinceId)
             throws TwitterException {
         return twitter.getMentions(getPaging(page, count, sinceId));
     }
@@ -464,7 +464,7 @@ public class TwitterConnector implements MuleContextAware {
     @Processor
     public ResponseList<Status> getRetweetedByMe(@Placement(group = "Pagination") @Default(value = "1") @Optional int page,
                                                  @Placement(group = "Pagination") @Default(value = "100") @Optional int count,
-                                                 @Placement(group = "Pagination") @Default(value = "-1") @Optional int sinceId)
+                                                 @Placement(group = "Pagination") @Default(value = "-1") @Optional long sinceId)
             throws TwitterException {
         return twitter.getRetweetedByMe(getPaging(page, count, sinceId));
     }
@@ -492,7 +492,7 @@ public class TwitterConnector implements MuleContextAware {
     @Processor
     public ResponseList<Status> getRetweetedToMe(@Placement(group = "Pagination") @Default(value = "1") @Optional int page,
                                                  @Placement(group = "Pagination") @Default(value = "100") @Optional int count,
-                                                 @Placement(group = "Pagination") @Default(value = "-1") @Optional int sinceId)
+                                                 @Placement(group = "Pagination") @Default(value = "-1") @Optional long sinceId)
             throws TwitterException {
         return twitter.getRetweetedToMe(getPaging(page, count, sinceId));
     }
@@ -519,7 +519,7 @@ public class TwitterConnector implements MuleContextAware {
     @Processor
     public ResponseList<Status> getRetweetsOfMe(@Placement(group = "Pagination") @Default(value = "1") @Optional int page,
                                                 @Placement(group = "Pagination") @Default(value = "100") @Optional int count,
-                                                @Placement(group = "Pagination") @Default(value = "-1") @Optional int sinceId)
+                                                @Placement(group = "Pagination") @Default(value = "-1") @Optional long sinceId)
             throws TwitterException {
         return twitter.getRetweetsOfMe(getPaging(page, count, sinceId));
     }
@@ -552,7 +552,7 @@ public class TwitterConnector implements MuleContextAware {
     public ResponseList<Status> getRetweetedToUserByScreenName(String screenName,
                                                                @Placement(group = "Pagination") @Default(value = "1") @Optional int page,
                                                                @Placement(group = "Pagination") @Default(value = "100") @Optional int count,
-                                                               @Placement(group = "Pagination") @Default(value = "-1") @Optional int sinceId)
+                                                               @Placement(group = "Pagination") @Default(value = "-1") @Optional long sinceId)
             throws TwitterException {
         return twitter.getRetweetedToUser(screenName, getPaging(page, count, sinceId));
     }
@@ -585,7 +585,7 @@ public class TwitterConnector implements MuleContextAware {
     public ResponseList<Status> getRetweetedToUserByUserId(long userId,
                                                            @Placement(group = "Pagination") @Default(value = "1") @Optional int page,
                                                            @Placement(group = "Pagination") @Default(value = "100") @Optional int count,
-                                                           @Placement(group = "Pagination") @Default(value = "-1") @Optional int sinceId)
+                                                           @Placement(group = "Pagination") @Default(value = "-1") @Optional long sinceId)
             throws TwitterException {
         return twitter.getRetweetedToUser(userId, getPaging(page, count, sinceId));
     }
@@ -618,7 +618,7 @@ public class TwitterConnector implements MuleContextAware {
     public ResponseList<Status> getRetweetedByUserByScreenName(String screenName,
                                                                @Placement(group = "Pagination") @Default(value = "1") @Optional int page,
                                                                @Placement(group = "Pagination") @Default(value = "100") @Optional int count,
-                                                               @Placement(group = "Pagination") @Default(value = "-1") @Optional int sinceId)
+                                                               @Placement(group = "Pagination") @Default(value = "-1") @Optional long sinceId)
             throws TwitterException {
         return twitter.getRetweetedByUser(screenName, getPaging(page, count, sinceId));
     }
@@ -651,7 +651,7 @@ public class TwitterConnector implements MuleContextAware {
     public ResponseList<Status> getRetweetedByUserByUserId(long userId,
                                                            @Placement(group = "Pagination") @Default(value = "1") @Optional int page,
                                                            @Placement(group = "Pagination") @Default(value = "100") @Optional int count,
-                                                           @Placement(group = "Pagination") @Default(value = "-1") @Optional int sinceId)
+                                                           @Placement(group = "Pagination") @Default(value = "-1") @Optional long sinceId)
             throws TwitterException {
         return twitter.getRetweetedByUser(userId, getPaging(page, count, sinceId));
     }
@@ -807,7 +807,7 @@ public class TwitterConnector implements MuleContextAware {
     public ResponseList<User> getRetweetedBy(long statusId,
                                              @Placement(group = "Pagination") @Default(value = "1") @Optional int page,
                                              @Placement(group = "Pagination") @Default(value = "100") @Optional int count,
-                                             @Placement(group = "Pagination") @Default(value = "-1") @Optional int sinceId)
+                                             @Placement(group = "Pagination") @Default(value = "-1") @Optional long sinceId)
             throws TwitterException {
         return twitter.getRetweetedBy(statusId, getPaging(page, count, sinceId));
     }
@@ -837,7 +837,7 @@ public class TwitterConnector implements MuleContextAware {
     public IDs getRetweetedByIDs(long statusId,
                                  @Placement(group = "Pagination") @Default(value = "1") @Optional int page,
                                  @Placement(group = "Pagination") @Default(value = "100") @Optional int count,
-                                 @Placement(group = "Pagination") @Default(value = "-1") @Optional int sinceId) throws TwitterException {
+                                 @Placement(group = "Pagination") @Default(value = "-1") @Optional long sinceId) throws TwitterException {
         return twitter.getRetweetedByIDs(statusId, getPaging(page, count, sinceId));
     }
